created auth.tsx; replaces index.tsx with static home page
